feat: adding climb up key
When using controlify, climb up doesn't work due to
some overwritting controlify does in the jump key.
This happens only with the gamepad mapping.
Having an optional climb up key map solves that problem.
I've also added a release climb key. When it is bound and you cling to a cliff, you say hanging even if you don't keep the key pressed. This works really well when you are a player with a controller.