feat: should surfing compatibility
When using a decoupled camera with shoulder surfing,
the actions taken when clinging are a bit messed up.
This is because it is confusing to climb up while hanging
if the camera is not following the direction you're looking at.
With this adaptation, we can force a coupled camera only
in this situation when it is decoupled