refactor: using future Shoulder Surfing feature
A proposal was sent to shoulder surfing to add a
plugin that allows othre mods to force a coupled
camera. This will be really useful to keep a
more compatible camera for some actions that may have
problems dealing with a decoupled one. The feature
is not available yet, but I'll let the branch ready
for it as an example