fix: accepting crawling while crouching
In my previous PR, I treated crawling x dodge disambiguation
to make crawling work when crouching, but it really didn't
work because crawl required the pose to be standing. I thought
it wasn't a problem because my previous tests worked, but I
can't really say why (maybe in 1.21.1 it worked, or
something was different in the code I didn't notice).
Allowing crawl to work with either Crouching and Standing
solves the issue